Fix parsing of wmctrl -lxp output lines
wmctrl actually seems to put two spaces either side of the wm_ctrl
column in the output (and one space either side of every other column).
Not being aware of this was causing the line parsing to fail for windows
that have spaces in their WM_CLASS, e.g. Hamster Time Tracker.